requirements.txt
numpy==2.3.5
opencv-python-headless==4.13.0.92
pillow==12.0.0
easyocr==1.7.2
rapidocr-onnxruntime==1.4.4
transformers==5.1.0
timm==1.0.24
nltk==3.9.2
tabulate==0.9.0
pandas==3.0.0
